flask/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
cffi==1.12.1
Click==7.0
cryptography==2.5
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mysql-connector-python==8.0.15
pkg-resources==0.0.0
protobuf==3.7.0
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==19.0.0
six==1.12.0
Werkzeug==0.14.1
